# openapi/plystra.v0.0.1.yaml
components:
    schemas:
        ApiAdminGrantMutationRequest:
            properties:
                expires_at:
                    format: date-time
                    nullable: true
                    type: string
                group_id:
                    type: string
                id:
                    type: string
                level:
                    type: string
                member_id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                permission_key:
                    type: string
                revoked_reason:
                    type: string
                space_id:
                    type: string
                status:
                    type: string
                user_id:
                    type: string
            type: object
...
        ApiOpenAPIAdminGrant:
            properties:
                created_at:
                    format: date-time
                    type: string
                deleted_at:
                    format: date-time
                    nullable: true
                    type: string
                expires_at:
                    format: date-time
                    nullable: true
                    type: string
                granted_by_member_id:
                    type: string
                granted_by_user_id:
                    type: string
                group_id:
                    type: string
                id:
                    type: string
                level:
                    example: space_admin
                    type: string
                member_id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                permission_key:
                    example: spaces:manage
                    type: string
                revoked_at:
                    format: date-time
                    nullable: true
                    type: string
                revoked_reason:
                    type: string
                space_id:
                    type: string
                status:
                    example: active
                    type: string
                updated_at:
                    format: date-time
                    type: string
                user_id:
                    type: string
            type: object
...
        ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIAdminGrant:
            properties:
                data:
                    $ref: '#/components/schemas/ApiOpenAPIAdminGrant'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIError:
            properties:
                audit_log_id:
                    example: audit_01hzyj6b6q8
                    type: string
                code:
                    example: VALIDATION_FAILED
                    type: string
                deny_code:
                    example: SCOPE_OUT_OF_BOUNDS
                    type: string
                details: {}
                message:
                    example: Request body is invalid JSON.
                    type: string
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
                trace_id:
                    example: trace_01hzyj6b6q8
                    type: string
            type: object
        ApiOpenAPIErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ApiOpenAPIError'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIAdminGrant:
            properties:
                data:
                    items:
                        $ref: '#/components/schemas/ApiOpenAPIAdminGrant'
                    nullable: true
                    type: array
                pagination:
                    $ref: '#/components/schemas/ApiOpenAPIPagination'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIPagination:
            properties:
                cursor:
                    nullable: true
                    type: string
                has_more:
                    example: false
                    type: boolean
                limit:
                    example: 50
                    type: integer
            type: object
...
    securitySchemes:
        ApiKeyAuth:
            description: Scoped API key for server-to-server Core calls.
            in: header
            name: X-Plystra-API-Key
            type: apiKey
        BearerAuth:
            bearerFormat: opaque Plystra access token
            description: Use the access_token returned by POST /api/v1/auth/login or POST /api/v1/auth/refresh.
            scheme: bearer
            type: http
        MetricsTokenAuth:
            description: Dedicated metrics token when METRICS_TOKEN or PLYSTRA_METRICS_TOKEN is configured.
            in: header
            name: X-Plystra-Metrics-Token
            type: apiKey
info:
    description: Self-hosted Plystra Core API. Stable surfaces cover identity, authorization, resource registry, audit, admin control-plane, and scoped API keys; plugin, template, and Data Console routes are preview metadata surfaces.
    title: Plystra Core API
    version: 0.0.1
openapi: 3.0.3
paths:
...
    /api/v1/admin/grants:
        get:
            operationId: listAdminGrants
            parameters:
                - description: Maximum number of rows to return.
                  in: query
                  name: limit
                  schema:
                    description: Maximum number of rows to return.
                    maximum: 200
                    minimum: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIAdminGrant'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List admin grants
            tags:
                - Admin
        post:
            operationId: createAdminGrant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiAdminGrantMutationRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIAdminGrant'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Create an admin grant
            tags:
                - Admin
    /api/v1/admin/grants/{admin_grant_id}:
        get:
            operationId: getAdminGrant
            parameters:
                - in: path
                  name: admin_grant_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIAdminGrant'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Get an admin grant
            tags:
                - Admin
    /api/v1/admin/grants/{admin_grant_id}/revoke:
        post:
            operationId: revokeAdminGrant
            parameters:
                - in: path
                  name: admin_grant_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiAdminGrantMutationRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIAdminGrant'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Revoke an admin grant
            tags:
                - Admin
...
servers:
    - url: http://localhost:8080
...